+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [Unreleased]
+
+### Added
+
+### Changed
+- Periodically refresh "Now playing" song to Last.fm while playing (Last.fm forgets the song after 4 minutes which was an issue on longer songs or when the song was paused)
+
+### Fixed
+
 ## 3.0.0 - 2026-02-23
 
 ### Breaking changes